requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
groq>=0.4.0

# Optional dependencies for enhanced functionality
# Install with: pip install -r requirements-optional.txt
# - chat-exporter>=2.8.0 (for advanced transcript formatting)
# - boto3>=1.26.0 (for S3 transcript storage)
